Feature/column2

+++ src/main/java/com/passion/teampassiontrelloproject/column/controller/ColumnController.java
@@ -22,31 +22,32 @@ public class ColumnController {
     private final ColumnsServiceImpl columnsService;
 
     @PostMapping("/columns")
-    public ResponseEntity<ColumnsResponseDto> createColumns(@AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody ColumnsResponseDto requestDto) {
-        ColumnsResponseDto result= columnsService.createColumns(requestDto, userDetails.getUser());
-
+    public ResponseEntity<ColumnsResponseDto> createColumns(@AuthenticationPrincipal UserDetailsImpl userDetails, @RequestBody ColumnsRequestDto requestDto) {
+        ColumnsResponseDto result = columnsService.createColumns(requestDto, userDetails.getUser());
         return ResponseEntity.status(HttpStatus.CREATED).body(result);
     }
 
     @PutMapping("/columns/{id}")
     public ResponseEntity<ApiResponseDto> updateColumns(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable Long id, @RequestBody ColumnsRequestDto requestDto) {
-        try {
-            Columns columns = columnsService.findColumns(id);
-            ColumnsResponseDto result = columnsService.updateColumns(columns, requestDto, userDetails.getUser());
-            return ResponseEntity.ok().body(result);
-        } catch (RejectedExecutionException e) {
-            return ResponseEntity.badRequest().body(new ApiResponseDto("작성자만 수정 할 수 있습니다.", HttpStatus.BAD_REQUEST.value()));
+        Columns columns = columnsService.findColumns(id);
+
+        if (!columns.getUser().getId().equals(userDetails.getUser().getId())) {
+            throw new IllegalArgumentException("작성자만 수정 할 수 있습니다.");
         }
+
+        ColumnsResponseDto result = columnsService.updateColumns(columns, requestDto, userDetails.getUser());
+        return ResponseEntity.ok().body(result);
     }
 
     @DeleteMapping("/columns/{id}")
     public ResponseEntity<ApiResponseDto> deleteColumns(@AuthenticationPrincipal UserDetailsImpl userDetails, @PathVariable Long id) {
-        try {
-            Columns columns = columnsService.findColumns(id);
-            columnsService.deleteColumns(columns, userDetails.getUser());
-            return ResponseEntity.ok().body(new ApiResponseDto("댓글 삭제 성공", HttpStatus.OK.value()));
-        } catch (RejectedExecutionException e) {
-            return ResponseEntity.badRequest().body(new ApiResponseDto("작성자만 삭제 할 수 있습니다.", HttpStatus.BAD_REQUEST.value()));
+        Columns columns = columnsService.findColumns(id);
+
+        if (!columns.getUser().getId().equals(userDetails.getUser().getId())) {
+            throw new IllegalArgumentException("작성자만 삭제 할 수 있습니다.");
         }
+
+        columnsService.deleteColumns(columns, userDetails.getUser());
+        return ResponseEntity.ok().body(new ApiResponseDto("컬럼 삭제 성공", HttpStatus.OK.value()));
     }
 }
